+++ docs/conf.py
@@ -24,9 +24,9 @@ copyright = '2018-2020, Lefteris Karapetsas'
 author = 'Lefteris Karapetsas'
 
 # The short X.Y version
-version = '1.1.1'
+version = '1.2.1'
 # The full version, including alpha/beta/rc tags
-release = '1.1.1'
+release = '1.2.1'
 
 
 # -- General configuration ---------------------------------------------------