+++ electron-app/src/utils/premium.ts
@@ -1,6 +1,5 @@
-import { VueConstructor } from 'vue';
+import Vue, { VueConstructor } from 'vue';
 import { api } from '@/services/rotkehlchen-api';
-import Vue from 'vue';
 import Chart from 'chart.js';
 import moment from 'moment';
 
@@ -10,25 +9,58 @@ export const setupPremium = () => {
   window.moment = moment;
 };
 
-export const PremiumStatistics = (): Promise<VueConstructor> => {
+function findComponents(): string[] {
+  return Object.getOwnPropertyNames(window).filter(value =>
+    value.startsWith('PremiumComponents')
+  );
+}
+
+async function loadComponents(): Promise<string[]> {
   return new Promise(async (resolve, reject) => {
+    let components = findComponents();
+    if (components.length > 0) {
+      resolve(components);
+      return;
+    }
+
     const result = await api.queryStatisticsRenderer();
     const script = document.createElement('script');
     script.text = result;
     document.head.appendChild(script);
-    const components = Object.getOwnPropertyNames(window).filter(value =>
-      value.startsWith('PremiumComponents')
-    );
+
+    components = findComponents();
+
     if (components.length === 0) {
       reject(new Error('There was no component loaded'));
+      return;
     }
-    // @ts-ignore
-    const library = window[components[0]];
-    Vue.use(library.install);
-    resolve(library.PremiumStatistics);
 
     script.addEventListener('error', reject);
+    resolve(components);
   });
+}
+
+async function loadLibrary() {
+  const [component] = await loadComponents();
+  // @ts-ignore
+  const library = window[component];
+  if (!library.installed) {
+    Vue.use(library.install);
+    library.installed = true;
+  }
+  return library;
+}
+
+export const PremiumStatistics = (): Promise<VueConstructor> => {
+  return new Promise(async resolve =>
+    resolve((await loadLibrary()).PremiumStatistics)
+  );
+};
+
+export const DsrMovementHistory = (): Promise<VueConstructor> => {
+  return new Promise(async resolve =>
+    resolve((await loadLibrary()).DsrMovementHistory)
+  );
 };
 
 declare global {