+++ rotkehlchen/api/v1/resources.py
@@ -3,8 +3,10 @@ from typing import Dict, List, Optional
 
 from flask import Blueprint, Response
 from flask_restful import Resource
+from marshmallow.utils import missing
 from typing_extensions import Literal
-from webargs.flaskparser import use_kwargs
+from webargs.flaskparser import parser, use_kwargs
+from webargs.multidictproxy import MultiDictProxy
 
 from rotkehlchen.api.v1.encoding import (
     AllBalancesQuerySchema,
@@ -62,6 +64,55 @@ from rotkehlchen.typing import (
 )
 
 
+def _combine_data_and_view_args(data, view_args, schema) -> MultiDictProxy:
+    if view_args is not missing:
+        if data == {}:
+            data = MultiDictProxy(view_args, schema)
+        else:
+            all_data = data.to_dict() if isinstance(data, MultiDictProxy) else data
+            for key, value in view_args.items():
+                all_data[key] = value
+            data = MultiDictProxy(all_data, schema)
+    return data
+
+
+@parser.location_loader('json_and_view_args')
+def load_json_viewargs_data(request, schema):
+    """Load data from a request accepting either json or view_args encoded data"""
+    view_args = parser.load_view_args(request, schema)
+    data = parser.load_json(request, schema)
+    if data is missing:
+        return data
+
+    data = _combine_data_and_view_args(data, view_args, schema)
+    return data
+
+
+@parser.location_loader('json_and_query')
+def load_json_query_data(request, schema):
+    """Load data from a request accepting either json or query encoded data"""
+    data = parser.load_json(request, schema)
+    if data is not missing:
+        return data
+    return parser.load_querystring(request, schema)
+
+
+@parser.location_loader('json_and_query_and_view_args')
+def load_json_query_viewargs_data(request, schema):
+    """Load data from a request accepting either json or querystring or view_args encoded data"""
+    view_args = parser.load_view_args(request, schema)
+    # Get data either from json or from querystring
+    data = parser.load_json(request, schema)
+    if data is missing:
+        data = parser.load_querystring(request, schema)
+
+    if data is missing:
+        return data
+
+    data = _combine_data_and_view_args(data, view_args, schema)
+    return data
+
+
 def create_blueprint():
     # Take a look at this SO question on hints how to organize versioned
     # API with flask:
@@ -79,7 +130,7 @@ class SettingsResource(BaseResource):
 
     put_schema = ModifiableSettingsSchema()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             premium_should_sync: Optional[bool],
@@ -119,7 +170,7 @@ class AsyncTasksResource(BaseResource):
 
     get_schema = AsyncTasksQuerySchema()
 
-    @use_kwargs(get_schema, locations=('view_args',))
+    @use_kwargs(get_schema, location='view_args')
     def get(self, task_id: Optional[int]) -> Response:
         return self.rest_api.query_tasks_outcome(task_id=task_id)
 
@@ -128,7 +179,7 @@ class FiatExchangeRatesResource(BaseResource):
 
     get_schema = FiatExchangeRatesSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, currencies: Optional[List[Asset]]) -> Response:
         return self.rest_api.get_fiat_exchange_rates(currencies=currencies)
 
@@ -141,7 +192,7 @@ class ExchangesResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_exchanges()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             name: str,
@@ -151,7 +202,7 @@ class ExchangesResource(BaseResource):
     ) -> Response:
         return self.rest_api.setup_exchange(name, api_key, api_secret, passphrase)
 
-    @use_kwargs(delete_schema, locations=('json',))
+    @use_kwargs(delete_schema, location='json')
     def delete(self, name: str) -> Response:
         return self.rest_api.remove_exchange(name=name)
 
@@ -164,14 +215,14 @@ class ExternalServicesResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_external_services()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             services: List[ExternalServiceApiCredentials],
     ) -> Response:
         return self.rest_api.add_external_services(services=services)
 
-    @use_kwargs(delete_schema, locations=('json',))
+    @use_kwargs(delete_schema, location='json')
     def delete(self, services: List[ExternalService]) -> Response:
         return self.rest_api.delete_external_services(services=services)
 
@@ -180,7 +231,7 @@ class AllBalancesResource(BaseResource):
 
     get_schema = AllBalancesQuerySchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, save_data: bool, async_query: bool, ignore_cache: bool) -> Response:
         return self.rest_api.query_all_balances(
             save_data=save_data,
@@ -193,7 +244,7 @@ class ExchangeBalancesResource(BaseResource):
 
     get_schema = ExchangeBalanceQuerySchema()
 
-    @use_kwargs(get_schema, locations=('json', 'view_args', 'query'))
+    @use_kwargs(get_schema, location='json_and_query_and_view_args')
     def get(self, name: Optional[str], async_query: bool, ignore_cache: bool) -> Response:
         return self.rest_api.query_exchange_balances(
             name=name,
@@ -212,7 +263,7 @@ class ExchangeTradesResource(BaseResource):
 
     get_schema = ExchangeTradesQuerySchema()
 
-    @use_kwargs(get_schema, locations=('json', 'view_args', 'query'))
+    @use_kwargs(get_schema, location='json_and_query_and_view_args')
     def get(
             self,
             name: Optional[str],
@@ -232,7 +283,7 @@ class BlockchainBalancesResource(BaseResource):
 
     get_schema = BlockchainBalanceQuerySchema()
 
-    @use_kwargs(get_schema, locations=('json', 'view_args', 'query'))
+    @use_kwargs(get_schema, location='json_and_query_and_view_args')
     def get(
             self,
             blockchain: Optional[SupportedBlockchain],
@@ -253,7 +304,7 @@ class FiatBalancesResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.query_fiat_balances()
 
-    @use_kwargs(patch_schema, locations=('json',))
+    @use_kwargs(patch_schema, location='json')
     def patch(self, balances: Dict[Asset, AssetAmount]) -> Response:
         # The passsed assets are guaranteed to be FIAT assets thanks to marshmallow
         return self.rest_api.set_fiat_balances(balances)
@@ -266,7 +317,7 @@ class TradesResource(BaseResource):
     patch_schema = TradePatchSchema()
     delete_schema = TradeDeleteSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(
             self,
             from_timestamp: Timestamp,
@@ -279,7 +330,7 @@ class TradesResource(BaseResource):
             location=location,
         )
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             timestamp: Timestamp,
@@ -306,7 +357,7 @@ class TradesResource(BaseResource):
             notes=notes,
         )
 
-    @use_kwargs(patch_schema, locations=('json',))
+    @use_kwargs(patch_schema, location='json')
     def patch(
             self,
             trade_id: str,
@@ -335,7 +386,7 @@ class TradesResource(BaseResource):
             notes=notes,
         )
 
-    @use_kwargs(delete_schema, locations=('json',))
+    @use_kwargs(delete_schema, location='json')
     def delete(self, trade_id: str) -> Response:
         return self.rest_api.delete_trade(trade_id=trade_id)
 
@@ -349,7 +400,7 @@ class TagsResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_tags()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             name: str,
@@ -364,7 +415,7 @@ class TagsResource(BaseResource):
             foreground_color=foreground_color,
         )
 
-    @use_kwargs(patch_schema, locations=('json',))
+    @use_kwargs(patch_schema, location='json')
     def patch(
             self,
             name: str,
@@ -379,7 +430,7 @@ class TagsResource(BaseResource):
             foreground_color=foreground_color,
         )
 
-    @use_kwargs(delete_schema, locations=('json',))
+    @use_kwargs(delete_schema, location='json')
     def delete(self, name: str) -> Response:
         return self.rest_api.delete_tag(name=name)
 
@@ -391,7 +442,7 @@ class UsersResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_users()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(
             self,
             name: str,
@@ -410,7 +461,7 @@ class UsersResource(BaseResource):
 class UsersByNameResource(BaseResource):
     patch_schema = UserActionSchema()
 
-    @use_kwargs(patch_schema, locations=('json', 'view_args'))
+    @use_kwargs(patch_schema, location='json_and_view_args')
     def patch(
             self,
             action: Optional[str],
@@ -448,7 +499,7 @@ class StatisticsAssetBalanceResource(BaseResource):
 
     get_schema = StatisticsAssetBalanceSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'view_args', 'query'))
+    @use_kwargs(get_schema, location='json_and_query_and_view_args')
     def get(
             self,
             asset: Asset,
@@ -466,7 +517,7 @@ class StatisticsValueDistributionResource(BaseResource):
 
     get_schema = StatisticsValueDistributionSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, distribution_by: str) -> Response:
         return self.rest_api.query_value_distribution_data(
             distribution_by=distribution_by,
@@ -489,7 +540,7 @@ class HistoryProcessingResource(BaseResource):
 
     get_schema = HistoryProcessingSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(
             self,
             from_timestamp: Timestamp,
@@ -507,7 +558,7 @@ class HistoryExportingResource(BaseResource):
 
     get_schema = HistoryExportingSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, directory_path: Path) -> Response:
         return self.rest_api.export_processed_history_csv(directory_path=directory_path)
 
@@ -525,7 +576,7 @@ class EthereumTokensResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_eth_tokens()
 
-    @use_kwargs(modify_schema, locations=('json',))
+    @use_kwargs(modify_schema, location='json')
     def put(
             self,
             eth_tokens: List[EthereumToken],
@@ -533,7 +584,7 @@ class EthereumTokensResource(BaseResource):
     ) -> Response:
         return self.rest_api.add_owned_eth_tokens(tokens=eth_tokens, async_query=async_query)
 
-    @use_kwargs(modify_schema, locations=('json',))
+    @use_kwargs(modify_schema, location='json')
     def delete(
             self,
             eth_tokens: List[EthereumToken],
@@ -549,11 +600,11 @@ class BlockchainsAccountsResource(BaseResource):
     patch_schema = BlockchainAccountsPatchSchema()
     delete_schema = BlockchainAccountsDeleteSchema()
 
-    @use_kwargs(get_schema, locations=('view_args',))
+    @use_kwargs(get_schema, location='view_args')
     def get(self, blockchain: SupportedBlockchain) -> Response:
         return self.rest_api.get_blockchain_accounts(blockchain)
 
-    @use_kwargs(put_schema, locations=('json', 'view_args'))
+    @use_kwargs(put_schema, location='json_and_view_args')
     def put(
             self,
             blockchain: SupportedBlockchain,
@@ -573,7 +624,7 @@ class BlockchainsAccountsResource(BaseResource):
             async_query=async_query,
         )
 
-    @use_kwargs(patch_schema, locations=('json', 'view_args'))
+    @use_kwargs(patch_schema, location='json_and_view_args')
     def patch(
             self,
             blockchain: SupportedBlockchain,
@@ -591,7 +642,7 @@ class BlockchainsAccountsResource(BaseResource):
             account_data=account_data,
         )
 
-    @use_kwargs(delete_schema, locations=('json', 'view_args'))
+    @use_kwargs(delete_schema, location='json_and_view_args')
     def delete(
             self,
             blockchain: SupportedBlockchain,
@@ -612,11 +663,11 @@ class IgnoredAssetsResource(BaseResource):
     def get(self) -> Response:
         return self.rest_api.get_ignored_assets()
 
-    @use_kwargs(modify_schema, locations=('json',))
+    @use_kwargs(modify_schema, location='json')
     def put(self, assets: List[Asset]) -> Response:
         return self.rest_api.add_ignored_assets(assets=assets)
 
-    @use_kwargs(modify_schema, locations=('json',))
+    @use_kwargs(modify_schema, location='json')
     def delete(self, assets: List[Asset]) -> Response:
         return self.rest_api.remove_ignored_assets(assets=assets)
 
@@ -631,7 +682,7 @@ class DataImportResource(BaseResource):
 
     put_schema = DataImportSchema()
 
-    @use_kwargs(put_schema, locations=('json',))
+    @use_kwargs(put_schema, location='json')
     def put(self, source: Literal['cointracking.info'], filepath: Path) -> None:
         return self.rest_api.import_data(source=source, filepath=filepath)
 
@@ -640,7 +691,7 @@ class MakerDAODSRBalanceResource(BaseResource):
 
     get_schema = AsyncQueryArgumentSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, async_query: bool) -> Response:
         return self.rest_api.get_makerdao_dsr_balance(async_query)
 
@@ -649,6 +700,6 @@ class MakerDAODSRHistoryResource(BaseResource):
 
     get_schema = AsyncQueryArgumentSchema()
 
-    @use_kwargs(get_schema, locations=('json', 'query'))
+    @use_kwargs(get_schema, location='json_and_query')
     def get(self, async_query: bool) -> Response:
         return self.rest_api.get_makerdao_dsr_history(async_query)