+++ rotkehlchen/tests/api/test_external_services.py
@@ -151,7 +151,7 @@ def test_add_external_services_errors(rotkehlchen_api_server):
     )
     assert_error_response(
         response=response,
-        contained_in_msg="services': {0: {'_schema': ['Invalid input type",
+        contained_in_msg="'services': ['Not a valid list.'",
         status_code=HTTPStatus.BAD_REQUEST,
     )
 
@@ -248,7 +248,7 @@ def test_remove_external_services_errors(rotkehlchen_api_server):
     )
     assert_error_response(
         response=response,
-        contained_in_msg="'services': {0: ['External service name should be a string']",
+        contained_in_msg="'services': ['Not a valid list.'",
         status_code=HTTPStatus.BAD_REQUEST,
     )